🎨 refactor: move optional class to separate file
